+++ packages/modules/orders/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./application";
 export * from "./module";
 export * from "./order";
 export * from "./orders";